Day 1 of passion
I'm pissed off right now , 1000/- may be shit for others but it means a lot to me, this has nothing to do with family its about my respect for money and the other possibilities followed by butterfly effect on it. I sacrificed my possibility for not some joke. I have to make it worth one way or another.